+++ packages/spacecat-shared-data-access/test/unit/models/site.test.js
@@ -20,6 +20,30 @@ import { sleep } from '../util.js';
 const validData = {
   baseURL: 'https://www.example.com',
   deliveryType: 'aem_edge',
+  hlxConfig: {
+    rso: {
+      owner: 'some-owner',
+      site: 'some-site',
+      ref: 'main',
+    },
+    cdnProdHost: 'www.example.com',
+    code: {
+      owner: 'some-owner',
+      repo: 'some-repo',
+      source: {
+        type: 'github',
+        url: 'https://github.com/some-owner/some-repo',
+      },
+    },
+    content: {
+      contentBusId: '1234',
+      source: {
+        type: 'onedrive',
+        url: 'https://some-owner.sharepoint.com/:f:/r/sites/SomeFolder/Shared%20Documents/some-site/www',
+      },
+    },
+    hlxVersion: 5,
+  },
   organizationId: 'org123',
 };
 
@@ -33,10 +57,20 @@ describe('Site Model Tests', () => {
       expect(() => createSite({ ...validData, deliveryType: 'invalid' })).to.throw('Invalid delivery type: invalid');
     });
 
+    it('throws an error if hlxConfig is invalid', () => {
+      expect(() => createSite({ ...validData, hlxConfig: '1234' })).to.throw('HLX Config must be an object: 1234');
+    });
+
     it('creates a site object with valid baseURL', () => {
       const site = createSite({ ...validData });
       expect(site).to.be.an('object');
       expect(site.getBaseURL()).to.equal(validData.baseURL);
+    });
+
+    it('creates a site without hlxConfig', () => {
+      const site = createSite({ ...validData, hlxConfig: undefined });
+      expect(site).to.be.an('object');
+      expect(site.getHlxConfig()).to.deep.equal({});
     });
 
     it('creates a site with default organization id', () => {
@@ -62,6 +96,33 @@ describe('Site Model Tests', () => {
       expect(siteConfig.getSlackConfig()).to.be.an('object');
       expect(siteConfig.getSlackMentions('type1')).to.deep.equal(['slackId1']);
       expect(siteConfig.getSlackMentions('type2')).to.deep.equal(['slackId2']);
+    });
+
+    it('creates a site with provided hlxConfig', () => {
+      const newHlxConfig = {
+        cdnProdHost: 'www.another-example.com',
+        code: {
+          owner: 'another-owner',
+          repo: 'another-repo',
+          source: {
+            type: 'github',
+            url: 'https://github.com/another-owner/another-repo',
+          },
+        },
+        content: {
+          contentBusId: '1234',
+          source: {
+            type: 'onedrive',
+            url: 'https://another-owner.sharepoint.com/:f:/r/sites/SomeFolder/Shared%20Documents/another-site/www',
+          },
+        },
+        hlxVersion: 5,
+      };
+      const site = createSite({ ...validData, hlxConfig: newHlxConfig });
+      const helixConfig = site.getHlxConfig();
+
+      expect(helixConfig).to.be.an('object');
+      expect(helixConfig).to.deep.equal(newHlxConfig);
     });
   });
 
@@ -123,6 +184,34 @@ describe('Site Model Tests', () => {
       expect(site.getGitHubURL()).to.equal(newGitHubURL);
     });
 
+    it('updates hlxConfig correctly', () => {
+      const newHlxConfig = {
+        cdnProdHost: 'www.another-example.com',
+        code: {
+          owner: 'another-owner',
+          repo: 'another-repo',
+          source: {
+            type: 'github',
+            url: 'https://github.com/another-owner/another-repo',
+          },
+        },
+        content: {
+          contentBusId: '1234',
+          source: {
+            type: 'onedrive',
+            url: 'https://another-owner.sharepoint.com/:f:/r/sites/SomeFolder/Shared%20Documents/another-site/www',
+          },
+        },
+        hlxVersion: 5,
+      };
+
+      site.updateHlxConfig(newHlxConfig);
+      const helixConfig = site.getHlxConfig();
+
+      expect(helixConfig).to.be.an('object');
+      expect(helixConfig).to.deep.equal(newHlxConfig);
+    });
+
     it('throws an error when updating with an invalid deliveryType', () => {
       expect(() => site.updateDeliveryType('invalid')).to.throw('Invalid delivery type: invalid');
     });
@@ -133,6 +222,10 @@ describe('Site Model Tests', () => {
 
     it('throws an error when updating with an invalid config', () => {
       expect(() => site.updateConfig('abcd')).to.throw('Config must be provided');
+    });
+
+    it('throws an error when updating with an invalid hlxConfig', () => {
+      expect(() => site.updateHlxConfig('abcd')).to.throw('HLX Config must be an object');
     });
 
     it('sets audits correctly', () => {